fix: format specifiers, type widths, mapper symbol
bklibopen.c:
- %zu → %"PRIu64" for int64_t end_sector/block_count
- Add #include <inttypes.h>
- size_t → uint64_t for start_sector/end_sector locals

bksrclib.c:
- size_t * → uint64_t * in get_new_header_range_and_offset params

bklibcreat.c, enclib.c:
- size_t → uint64_t for start_sector/end_sector

libplat/mapper.c:
- Define is_device_mapper_available in dispatcher (tentative for GNU, false for ISOC)
- Fixes undefined symbol under ISO C build

libplat/ISOC/init.c:
- Fix timer resolution spin-loop: && → || to handle sec-only timers

+++ backend/bklibcreat.c
@@ -71,7 +71,7 @@ void action_create(
    Data      data;
    uint8_t * aux_zone = NULL;
    uint8_t   master_key[HASHLEN];
-   size_t    start_sector, end_sector;
+   uint64_t   start_sector, end_sector;
    int       ret_target_level;
 
    fill_secure_random_bits(master_key, HASHLEN);
